added problems to blueprint
